requirements.txt
torch==2.0.1
transformers==4.35.0
openai-whisper==20231117 
pyaudio==0.2.14
pyyaml==6.0.1 
numpy==1.26.2 
requests==2.31.0 
loguru==0.7.2 
fastapi==0.115.0 
uvicorn==0.30.6 
gtts==2.5.3